Add insertion sort
I did this one with a fold but then searched to check my answer
and saw it being done with explicit recursion. I think the explicit
recursion is better for a warmup exercise since it bends my brain more.
I am going to commit that next but wanted to have a record of this.